+++ src/main/java/com/shiye/mir/service/impl/GetSeparatedServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class GetSeparatedServiceImpl implements GetSeparatedService {
     @Override
-    public void doSeparate(Long uid, boolean getBeat, boolean getVocal){
+    public void doSeparate(Integer uid, boolean getBeat, boolean getVocal){
         //1.从缓存中调出uid
 
         //2.调出redis中的文件
